+++ src/network/utils/error-model.h
@@ -388,22 +388,22 @@ class ListErrorModel : public ErrorModel
     /**
      * \return a copy of the underlying list
      */
-    std::list<uint32_t> GetList() const;
+    std::list<uint64_t> GetList() const;
     /**
      * \param packetlist The list of packet uids to error.
      *
      * This method overwrites any previously provided list.
      */
-    void SetList(const std::list<uint32_t>& packetlist);
+    void SetList(const std::list<uint64_t>& packetlist);
 
   private:
     bool DoCorrupt(Ptr<Packet> p) override;
     void DoReset() override;
 
     /// Typedef: packet Uid list
-    typedef std::list<uint32_t> PacketList;
+    typedef std::list<uint64_t> PacketList;
     /// Typedef: packet Uid list const iterator
-    typedef std::list<uint32_t>::const_iterator PacketListCI;
+    typedef std::list<uint64_t>::const_iterator PacketListCI;
 
     PacketList m_packetList; //!< container of Uid of packets to corrupt
 };
